apps/python-api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
protobuf==4.24.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
